+++ App.tsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import 'react-native-gesture-handler';
 import {NavigationContainer} from '@react-navigation/native';
-import Main from './srcs/screen/Main';
+import Main from './srcs/screens/Main';
 
 const App = () => {
   return (
     <NavigationContainer>
-      {/* <TabNavigator /> */}
       <Main />
     </NavigationContainer>
   );
