+++ babel.config.js
@@ -1,3 +1,16 @@
 module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
+  plugins: [
+    [
+      'module-resolver',
+      {
+        root: ['.'],
+        alias: {
+          '@screens': './srcs/screens',
+          '@common': './srcs/common',
+          '@srcs': './srcs',
+        },
+      },
+    ],
+  ],
 };
