+++ package.json
@@ -6,9 +6,9 @@
     "android": "react-native run-android",
     "ios": "react-native run-ios --simulator='iPhone 14'",
     "start": "react-native start",
-	"build-android": "cd android && ./gradlew clean build",
-	"build-ios": "cd ios && pod install",
-	"build": "yarn && yarn build-android && yarn build-ios",
+    "build-android": "cd android && ./gradlew clean build",
+    "build-ios": "cd ios && pod install",
+    "build": "yarn && yarn build-android && yarn build-ios",
     "test": "jest",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
@@ -48,6 +48,7 @@
     "@typescript-eslint/eslint-plugin": "^5.37.0",
     "@typescript-eslint/parser": "^5.37.0",
     "babel-jest": "^26.6.3",
+    "babel-plugin-module-resolver": "^4.1.0",
     "eslint": "^7.32.0",
     "jest": "^26.6.3",
     "metro-react-native-babel-preset": "^0.72.1",
