+++ srcs/common/type.ts
@@ -1,0 +1,3 @@
+import {Alarm} from 'react-native-simple-alarm/dist/Types';
+
+export type AlarmType = Alarm;
