+++ srcs/contexts/index.ts
@@ -1,3 +1,2 @@
-export * from './ToggleThemeContext';
 export * from './AutoFocusContext';
-export * from './ScrollEnabledContext';
+export * from './useAlarmUpdate';
