+++ srcs/libs/alarm/createAlarm.ts
@@ -1,5 +1,5 @@
 import {createAlarm as RNCreateAlarm} from 'react-native-simple-alarm';
-import {AlarmType} from '.';
+import {AlarmType} from '@common/type';
 
 export const createAlarm = async (props: AlarmType) => {
   const {active, date, message, snooze, soundName} = props;
