+++ srcs/libs/alarm/index.ts
@@ -2,4 +2,3 @@ export * from './createAlarm';
 export * from './deleteAlarm';
 export * from './editAlarm';
 export * from './getAlarm';
-export * from './alarmType';
