--- srcs/screen/AddAlarm.style.tsx
@@ -1,33 +1,0 @@
-import {StyleSheet} from 'react-native';
-import theme from '../styles/theme';
-
-export const styles = StyleSheet.create({
-  view: {
-    flex: 1,
-    flexDirection: 'column',
-    alignItems: 'center',
-    backgroundColor: theme.color.black,
-  },
-  tapListView: {
-    flexDirection: 'column',
-    alignSelf: 'center',
-    width: '90%',
-    marginLeft: 10,
-    marginRight: 10,
-    borderRadius: 5,
-  },
-  tapItemView: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    backgroundColor: 'transparent',
-    marginLeft: 12,
-    marginRight: 12,
-    margin: 10,
-  },
-  separator: {
-    height: 1,
-    marginHorizontal: 20,
-    backgroundColor: 'lightgray',
-  },
-});
