--- srcs/screen/AddAlarmDetail/index.ts
@@ -1,3 +1,0 @@
-export * from './Message';
-export * from './Repeat';
-export * from './Sound';
