+++ srcs/screens/add_alarm/components/AlarmDetail.styles.ts
@@ -1,5 +1,5 @@
 import {StyleSheet} from 'react-native';
-import theme from '../../styles/theme';
+import theme from '../../../common/theme';
 
 const color = theme.color;
