+++ srcs/screens/add_alarm/components/Sound.tsx
@@ -2,12 +2,12 @@ import React, {useLayoutEffect, useState} from 'react';
 import {FlatList, Pressable, Text, TouchableOpacity, View} from 'react-native';
 import {useNavigation} from '@react-navigation/native';
 import {StackNavigationProp} from '@react-navigation/stack';
-import {ModalStackParamList} from '../Main';
+import {ModalStackParamList} from '../../Main';
 import {styles} from './ButtonList.styles';
-import {SOUND} from '../../libs/alarm';
-import {useCreateAlarm, updateAction} from '../../contexts/CreateAlarm';
+import {SOUND} from '@common/constant';
+import {useCreateAlarm, updateAction} from '../../../contexts/CreateAlarm';
 import Icon from 'react-native-vector-icons/Ionicons';
-import theme from '../../styles/theme';
+import theme from '../../../common/theme';
 
 type songScreenProps = StackNavigationProp<ModalStackParamList, 'Song'>;
