+++ srcs/screens/home/components/ListItem.tsx
@@ -1,14 +1,15 @@
 import React from 'react';
-import {styles} from './ListItem.style';
 import {Swipeable, TouchableOpacity} from 'react-native-gesture-handler';
-import {Animated, Switch, Text, View} from 'react-native';
-import {AlarmType, deleteAlarmById, switchAlarmById} from '../libs/alarm';
+import {Animated, StyleSheet, Switch, Text, View} from 'react-native';
 import moment from 'moment';
-import {useAlarmUpdate} from '../contexts/useAlarmUpdate';
-import theme from '../styles/theme';
 import Icon from 'react-native-vector-icons/Ionicons';
+import theme from '@common/theme';
+import {AlarmType} from '@common/type';
+import {deleteAlarmById, switchAlarmById} from '@srcs/libs/alarm';
+import {useAlarmUpdate} from '../../../contexts';
+// import {useAlarmUpdate} from 'srcs/contexts';
 
-const ListItem = (props: AlarmType) => {
+export default function ListItem(props: AlarmType) {
   const {setUpdated} = useAlarmUpdate();
   const {oid, active, date, message} = props;
   return (
@@ -35,7 +36,7 @@ const ListItem = (props: AlarmType) => {
       </View>
     </Swipeable>
   );
-};
+}
 
 const renderRightActions = (
   progress: Animated.AnimatedInterpolation<any>,
@@ -56,4 +57,48 @@ const renderRightActions = (
   );
 };
 
-export default ListItem;
+const styles = StyleSheet.create({
+  itemView: {
+    display: 'flex',
+    flexDirection: 'column',
+    marginVertical: 8,
+    paddingTop: 4,
+    paddingBottom: 12,
+    paddingHorizontal: 16,
+    backgroundColor: theme.color.grey30,
+  },
+
+  timeView: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+  },
+  timeText: {
+    color: theme.color.white,
+    fontSize: theme.fontSize.xl,
+    fontFamily: 'NotoSansKR-Medium',
+  },
+  messageView: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  messageText: {
+    fontSize: theme.fontSize.sm,
+    color: theme.color.grey99,
+    fontFamily: 'NotoSansKR-Regular',
+  },
+
+  /* Swipeables */
+  rightActionView: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'flex-end',
+    justifyItems: 'center',
+    //todo: color: system error
+    borderWidth: 1,
+    backgroundColor: theme.color.primary,
+    marginVertical: 8,
+    marginHorizontal: 10,
+    padding: 16,
+  },
+});
