+++ tsconfig.json
@@ -4,6 +4,12 @@
     /* Visit https://aka.ms/tsconfig.json to read more about this file */
 
     /* Completeness */
-    "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
+    "skipLibCheck": true,                                 /* Skip type checking all .d.ts files. */
+	"baseUrl": ".",
+	"paths": {
+		"@screens/*": ["srcs/screens/*"],
+		"@common/*": ["srcs/common/*"],
+		"@srcs/*": ["srcs/*"]
+	}
   }
 }
